requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.21
yt-dlp>=2023.3.4
pydub>=0.25.1